tui: fix rendering, upgrade ratatui 0.30, add unified diff, split app.rs
Rendering / robustness:
- Route library timing logs through `log` (debug) instead of eprintln,
  which was painting over the alternate screen and corrupting the TUI.
- Upgrade ratatui 0.26 -> 0.30 and crossterm 0.27 -> 0.29.
- Handle terminal Resize (clear + full repaint) and only act on key
  Press events to avoid double-processing; exit cleanly on a closed
  output pipe (SIGPIPE) instead of panicking.

Diff usability:
- `v` now toggles disassembly/decompiled view (and the diffed kind in the
  split view); previously it did nothing outside the diff view.
- Functions present only in file 2 ("added") now show their v2 code in the
  single-pane views instead of a "not present" stub.
- Add a git-style unified diff view (toggle with `u`): a single column
  showing the full before/after with -/+ markers and old/new line numbers.

Structure:
- Split the 960-line tui/app.rs into app.rs (state), content.rs
  (content/diff rendering) and background.rs (task polling).
