+++ crates/hbc-decomp-cli/Cargo.toml
@@ -13,10 +13,13 @@ path = "src/main.rs"
 [dependencies]
 clap = { version = "4.5.0", features = ["derive"] }
 hbc-decomp = { path = "../hbc-decomp" }
-ratatui = "0.26.3"
-crossterm = "0.27.0"
+ratatui = "0.30"
+crossterm = "0.29"
 regex = "1.10"
 serde_json = "1.0"
 log = "0.4.29"
 env_logger = "0.11.9"
 similar = "2.7.0"
+
+[target.'cfg(unix)'.dependencies]
+libc = "0.2"
